Session control from URL input added

+++ src/Application/Actions/User/DeleteUserAction.php
@@ -13,7 +13,27 @@ class DeleteUserAction extends UserAction {
      */
     protected function action(): Response
     {
+        if (!isset($_SESSION)) session_start();
+        $name = $_SERVER["SERVER_NAME"];
+        $port = ':'.$_SERVER["SERVER_PORT"];
+        
+        if (!isset($_SESSION['id']))
+        {
+            $dest = "/error" ;
+            header("Location: http://$name$port$dest");
+            exit();
+        }
+
         $userId = (int) $this->resolveArg('id');
+
+        // only logged user and admin can delete user from URL 
+        if ($userId !== $_SESSION['id'] && $_SESSION['role'] !== "Admin")
+        {
+            $dest = "/unauthorized_access_error";
+            header("Location: http://$name$port$dest");
+            exit();
+        }
+
         $this->userRepository->delete($userId);
         $this->logger->info("User of id `${userId}` was deleted.");
